requirements.txt
# AI 地址智能标准化工具 — Python 依赖
# 安装命令：pip install -r requirements.txt

# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 环境变量加载
python-dotenv==1.0.1

# OpenAI 兼容客户端（同时支持 DeepSeek 和 Claude 中转站）
openai==1.40.0

# 异步 HTTP（用于高德地图联网验证）
httpx==0.27.0

# 数据校验
pydantic==2.8.0
